+++ lib/helpers.ts
@@ -1,15 +1,23 @@
-import { AnySelector, SelectorTypes } from "./selectors";
+import { AnySelector } from "./selectors";
 import { ListType, AnyControl } from "./controls";
 
 export interface Query {
     [prop: string]: Query | AnySelector | AnyControl;
 }
 
-export type GetResult<Q extends AnySelector | ListType<any> | Query> = {
-    0: Q extends AnySelector ? SelectorTypes[Q["type"]] : never;
-    1: Q extends ListType<infer T> ? GetResult<T>[] : never;
+export interface SelectorResultType {
+    text: string;
+    attr: string;
+    count: number;
+    html: string;
+    exists: boolean;
+}
+
+export type TypeOfQuery<Q extends AnySelector | ListType<any> | Query> = {
+    0: Q extends AnySelector ? SelectorResultType[Q["type"]] : never;
+    1: Q extends ListType<infer T> ? TypeOfQuery<T>[] : never;
     2: Q extends Query ? {
-        [prop in keyof Q]: GetResult<Q[prop]>
+        [prop in keyof Q]: TypeOfQuery<Q[prop]>
     }: never;
 }[Q extends AnySelector ? 0
     : Q extends ListType<any> ? 1